Rename Slack events to actions
