+++ api/views/submission.py
@@ -42,7 +42,7 @@ from api.models import Source, Submission, Transcription
 from api.pagination import StandardResultsSetPagination
 from api.serializers import SubmissionSerializer
 from api.slack import client as slack
-from api.slack.events import ask_about_removing_post
+from api.slack.actions import ask_about_removing_post
 from api.slack.utils import _send_transcription_to_slack
 from api.views.volunteer import VolunteerViewSet
 from authentication.models import BlossomUser
